Fix error on analytics cluster view
`user_id` in the `Cluster` model is an integer field and `user_id` of `Ratings` is a char field. That's why when you try to visit a page of cluster (for example `http://localhost:8000/analytics/cluster/3/`) Django was throwing an error because the code was comparing integers and strings. Now, integer list is converted to a string list first and strings are compared to strings.